move(cosmo): move funcs to private API
