+++ astropy/cosmology/funcs.py
@@ -1,0 +1,6 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Astropy cosmology funcs module."""
+
+__all__ = ["cosmology_equal", "z_at_value"]
+
+from ._src.funcs import cosmology_equal, z_at_value
